separated all the files so they make more sense
