+++ gui/game.py
@@ -1,0 +1,6 @@
+
+
+class Game:
+    def __init__(self,a,b):
+        self.a = a
+        self.b = b
